backend/requirements.txt
Django==2.1.3
google-cloud-language==1.1.0
aiml==0.9.1
django-cors-headers==2.4.0
django-dotenv==1.4.2
djangorestframework==3.9.0
